Revision according to comments
Replace putData by checking the actual coordinator table metadata

Switch back constant to private
